ファイルの存在判定を追加

+++ setup.sh
@@ -4,7 +4,11 @@ DOT_FILES=(.zshrc .vimrc .vim .tmux.conf oh-my-zsh .iterm2_shell_integration.zsh
 
 for file in ${DOT_FILES[@]}
 do
-  ln -s $HOME/dotfiles/$file $HOME/$file
+  if [ ! -e $HOME/$file ]; then
+    ln -s $HOME/dotfiles/$file $HOME/$file
+  fi
 done
 
-ln -s $HOME/dotfiles/dpoggi_my.zsh-theme $HOME/oh-my-zsh/themes/dpoggi_my.zsh-theme
+if [ ! -e $HOME/oh-my-zsh/themes/dpoggi_my.zsh-theme ]; then
+  ln -s $HOME/dotfiles/dpoggi_my.zsh-theme $HOME/oh-my-zsh/themes/dpoggi_my.zsh-theme
+fi
